clean up & implement spatial profiles
